backend/requirements.txt
Flask==2.0.1
flasgger==0.9.5
flask-cors==3.0.10
psycopg2-binary==2.9.3
werkzeug==2.0.1
requests==2.27.1
python-dotenv==0.19.2
gunicorn==20.1.0